requirements.txt
absl-py>=2.1.0
asgiref>=3.8.1
astunparse>=1.6.3
certifi>=2024.2.2
charset-normalizer>=3.3.2
Django>=5.0.6
flatbuffers>=24.3.25
gast>=0.5.4
google-pasta>=0.2.0
grpcio>=1.63.0
h5py>=3.11.0
idna>=3.7
keras>=3.3.3
libclang>=18.1.1
Markdown>=3.6
markdown-it-py>=3.0.0
MarkupSafe>=2.1.5
mdurl>=0.1.2
ml-dtypes>=0.3.2
namex>=0.0.8
numpy>=1.26.4
opt-einsum>=3.3.0
optree>=0.11.0
packaging>=24.0
pillow>=10.3.0
protobuf>=4.25.3
Pygments>=2.18.0
requests>=2.31.0
rich>=13.7.1
scipy>=1.13.0
six>=1.16.0
sqlparse>=0.5.0
tensorboard>=2.16.2
tensorboard-data-server>=0.7.2
tensorflow>=2.16.1
tensorflow-intel>=2.16.1
tensorflow-io-gcs-filesystem>=0.31.0
termcolor>=2.4.0
typing_extensions>=4.11.0
tzdata>=2024.1
urllib3>=2.2.1
Werkzeug>=3.0.3
wrapt>=1.16.0